requirements.txt
discord.py==2.3.2
discord.py==1.7.3
GitPython==3.1.32
GitPython==3.1.42
interception_python==0.0.1
keyboard==0.13.5
mss==9.0.1
numpy==1.25.2
numpy==1.22.1
opencv_python==4.8.0.76
opencv_python==4.5.5.62
opencv_python_headless==4.8.0.76
Pillow==10.0.0
Pillow==10.2.0
PyAutoGUI==0.9.54
PyAutoGUI==0.9.53
pygame==2.5.1
pyotp==2.9.0
pytz==2023.3
pywin32==306
